requirements.txt
streamlit
pandas
google-generativeai
duckduckgo-search
matplotlib
seaborn